// approoveit-client/src/shared/types.ts
export interface Client {
  id: string;
  email: string;
  name: string;
  companyName?: string;
  profilePictureUrl?: string;
  createdAt: string;
  lastUsedAt?: string;
  submissionCount: number;
}